+++ jfreereport/source/com/jrefinery/report/targets/pageable/operations/LeftAlignment.java
@@ -28,7 +28,7 @@
  * Original Author:  Thomas Morgner;
  * Contributor(s):   David Gilbert (for Simba Management Limited);
  *
- * $Id: LeftAlignment.java,v 1.2 2003/02/02 23:43:51 taqua Exp $
+ * $Id: LeftAlignment.java,v 1.8 2003/02/07 22:40:42 taqua Exp $
  *
  * Changes
  * -------
@@ -74,11 +74,11 @@ public class LeftAlignment extends HorizontalBoundsAlignment
       throw new NullPointerException("Inner Bound must not be null");
     }
     inner = referenceBounds.createIntersection(inner);
-    double x = referenceBounds.getX();
-    double y = inner.getY();
-    double w = Math.min (inner.getWidth(), referenceBounds.getWidth());
-    double h = Math.min (inner.getHeight(), referenceBounds.getHeight());
+    float x = (float) referenceBounds.getX();
+    float y = (float) inner.getY();
+    float w = (float) Math.min (inner.getWidth(), referenceBounds.getWidth());
+    float h = (float) Math.min (inner.getHeight(), referenceBounds.getHeight());
 
-    return new Rectangle2D.Double(x, y, w, h);
+    return new Rectangle2D.Float(x, y, w, h);
   }
 }
